Give Verilator Hardware Model / Build and Test job a unique name
Unfortunately, Github doesn't always include the workflow name when
showing jobs in the UI, making it hard to tell the difference between
the "Build and Test" job in the "Build and Test" workflow and the "Build
and Test" job in the "Verilator Hardware Model" workflow.

So rename it to "Verilator Build and Test".