emu: Make InstrTracer a value type instead of a reference type.
This allows the caller of Cpu::step() more flexibility; in particular,
they can create an Option<Box<InstrTracer>> and convert it to
Option<&mut InstrTracer> with Option::as_deref_mut(). Also, the
reference lifetime and closure lifetime can be specified separately,
which is sometimes necessary to make the borrow checker happy.