Switch build-test runner from ubuntu 22.04 to 20.04.
This version of Ubuntu has more in common with the older Linux
distributions found at some corporations. For example, it uses openssl
1.1 instead of 3.0. By running this workflow with an older distribution,
it is less likely we accidentally break these users' build.

I didn't change the verilator workflow to 20.04, as it seems to having
severe performance problems under 20.04 that I don't understand.