Remove unnecessary deps from caliptra-image-* crates (#88)
Removes the bespoke version of clap, and the getset dependency. Transitively this removes 27 crates from Cargo.lock.

As they are used to generate official firmware images, the caliptra-image-* dependencies will eventually need to run in a trusted build environment where dependencies are vetted. Reducing the number of dependencies will reduce the code-review effort.