emu: ready_for_fw_cb should be called immediately.
Previously, special-case logic in the emulator library locked the
mailbox, then called ready_for_fw_cb 1000 cycles later, with no way for
users to opt-out of this logic.

In this commit, ready_for_fw_cb is called immediately, and provided an
API for scheduling mailbox actions in the future. The emulator app uses
this API to maintain the previous behavior, but other users don't have
to.