emu: Make CSR hacks in soc_ifc compatible with fix for #129
When flowstatus.[il]devid_csr_ready fields are set, the boot ROM enters
a special mode that signs a ldevid/idevid CSR and places it in the
mailbox. The emulator's soc_ifc peripheral has some hacks in the
implementation that check for this, and download the CSR from the
mailbox into a file "/tmp/caliptra_?devid_cert.der". Unfortunately, this
logic was broken by #132, as I didn't realize this hack exited at the
time. To fix:

  * The hack should look at mbox_fsm_ps field instead of status field to
    detect an incoming message from the Caliptra microcontroller.

  * The hack should set the status field to signal the microcontroller
    that it has read the incoming message