ROM: Don't set mailbox status during CSR upload.
The makefile passes --req-idevid-csr to the emulator app, which sets the
idevid_csr_ready fields in soc_ifc. When these are set, the boot ROM
enters a special mode that signs a ldevid/idevid CSR and places it in
the mailbox. When fixing #129 I didn't realize that this logic existed,
and didn't update it with the correct behavior.

When writing the CSR to the mailbox, the firmware MUST NOT set the
status. According to the integration guide, status should only be set by
the receiver of the command (the SoC). For more details, see #129.