HwModel: Remove test_firmware_upload
This test is incompatible with the latest RTL (senders can no longer
read their request from the FIFO).  This test will be replaced by a more
realistic SoC->uC mailbox test in a future commit after the RTL is
updated. Bugs in the current RTL prevent this test from being included
with this commit.