[feat] Remove message retrieval capability from keyVault for ECC384 signature operation (#180)
This change removes the capability for the ECC384 driver sign api to retrieve the message from the keyvault.
This is done to account for a similar change in the RTL (https://github.com/chipsalliance/caliptra-rtl/commit/d88fc6c2039fdcef9904d93f96e9d78e4c3accd1). Currently, there is no scenario in the ROM
that uses this capability.