ureg: Generate peripheral instance type.
In safe code, holders of an instance of this type are guaranteed
exclusive access to this peripheral.

Once all the drivers have been migrated, the old mechanism to get a
RegisterBlock from arbitrary safe-code will be removed, making it
impossible for safe code to manipulate a peripheral behind the back of a
driver.