Migrate Mailbox peripheral to new ureg instance type.
This commit is a bit trickier, as we need to borrow the mailbox
peripheral from the RomEnv during the mailbox transaction, which
prevents us from passing the RomEnv around while the mailbox transaction
is in progress (since the mailbox field in is use). To work around this,
we borrow the drivers we need from RomEnv and store the references in a
separate RomImageVerificationEnv, which we can use while the mailbox
transaction in progress.

Another approach to consider is removing the mailbox field from the
RomEnv struct.