Update rust toolchain from 1.68 to 1.70 (-2076 bytes).
Set the desired toolchain in rust-toolchain.toml, so everyone uses the
same toolchain by default, and we can remove the explicit risv32imc
target install steps.

The improved risc-v codegen in this release reduces the size of the ROM
by 2076 bytes.

As part of this change, we can remove the cargo index cache, as in 1.70
cargo switched to the "sparse index protocol":

https://github.com/rust-lang/rust/blob/642c92e63008ffb49f6ad8344e07bfa7d5b0d9bb/RELEASES.md?plain=1#L101