caliptra-test: Add common crypto operations.
To ensure that private keys are derived correctly, the integration tests
need to reproduce some cryptographic operations performed by the
caliptra hardware and firmware. To increase confidence that the same
mistake isn't made in the tests and the firmware, these are implemented
independently of the rest of the Caliptra codebase.