Update to the latest dev-integrate caliptra-rtl.
Also, allow sha384acc KAT to run when already locked.

With the latest RTL, the sha384acc peripheral starts in the
locked-for-uc state, to prevent the SoC from performing cryptographic
operations before the KAT has executed. try_start_operation() should
detect this state, and still allow the operation to be started.

Also, if the KAT fails, don't unlock the peripheral, since we can't
allow the SoC to use it.

Also, update various tests to unlock sha512acc before accessing it from
the SoC.

Also, update sw-emulator to lock the sha384acc peripheral at startup,
similar to the latest RTL.