libcaliptra
Derived from the api example in hw-model/c-binding/examples/api, this
is the initial skeleton of a more generic client API for Caliptra,
intended to abstract away the details of the implementation.

Client applications will include caliptra_api.h and implementations
will provide definitions to the functions declared in caliptra_if.h,
check the README.md for more information.