Add support for integration with the model
This adds a generic reference to demonstrate client application use of the
Caliptra C API, and an example implementation of an interface, in this case
connecting the generic reference application to the hardware model.

Build happens from within the examples/hwmodel/ directory. Run `make` and then
run `./hwmodel` to execute the example.

This also removes the more basic if_test directory as it is now redundant