CI speedup: Try to avoid invalidating cargo cache.
When compiling for the host, try to use the same RUSTFLAGS everywhere to
avoid invalidating the cargo cache.

Also, in the hw-model c binding tests, don't change the target
directory, so we can use the cached artifacts built earlier in the
workflow.

Together, these changes should reduce the total time from ~16 minutes to
~8 minutes.