Add ctx-based sha384 driver impl
Add a set of APIs to the sha384 driver to avoid needing a persistent
reference to the driver.

These APIs hold all state in a context struct which is then passed the
the sha384 driver to operate on it. This greatly simplifies
lifetime/borrow management of long-running hash operations. It is also
safe to have multiple active contexts at the same time (assuming
single-threaded execution).

This PR leaves the existing set of op-based APIs in place to avoid
changing ROM. This new set of APIs is primarily intended to be used in
runtime firmware.