Move DPE submodule to root
If (1) Crate A is in a workspace and (2) Crate B is in a subdirectory of a
crate A, then Crate B cannot be excluded from the workspace. This causes
some problems when trying to build DPE from within the workspace.

Move the DPE submodule out from under runtime so that it can be
excluded.