In sha drivers, take digest during finalize instead of init
Currently, the sha drivers take a digest reference during init, and then
don't use it until finalize is called. This can be complicated for
managing the lifetime of the digest structure. RT would like to be able
to hold a sha context which contains a SHA op and a digest; but the
lifetime management for doing this is very tricky.

Instead, take the digest as input to finalize and write the state to it
then.

This does not change any functionality of the SHA drivers, only the
interface.