ROM should not print out the FHT.
This does not actually do anything today because the ROM disables the
'emu' future on the caliptra-common crate, turning
caliptra-common::cprintln!() into a no-op. In a future commit, the FHT
logic will be moved to caliptra-drivers, where the rom enables the "emu"
feature for uart builds, and we don't want a size regression when that
happens.