Replace gen iter w/ fixed 12 and 16 words methods
Since the Trng driver will wrap the CSRNG and always request 12 words,
it doesn't make sense to have an iterator that returns a variable number
of words. No one else will be calling into the CSRNG directly.

The `generate16` method is for NIST test vectors, which requests 4
blocks of 4 words each.