[fuzzing] Fix x509 fuzzing target to match its seed corpus
The corpus is seeded with TBS blobs, so parse the input as TBS first. We
assume the fuzzer will generally follow the seed corpus for available
bytes, then generate other data, which will be parsed as the
'signature.'

Note that this has no effect on the current findings, the target panics
before and after. It is only intended to help the fuzzer make better use
of the seed corpus.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>