Activate authentication via libcaliptra (#651)
* Activate authentication via libcaliptra

This adds a demonstration of signature setup in caliptra fuses of firmware
hashes. Additionally, it reorganizes some of the headers and adds types and
structs for the caliptra firmware image to ease access to the public key
object within.

On startup, the demo will load the firmware and extract the vendor and
owner public keys. It will then generate a sha384 digest and program
them into the fuses - right as ROM is loaded.

This replicates behavior seen in the smoke test (test/tests/smoke_test.rs).

* Added fixed register values from the smoke test to enable auth.
* Added support for OpenSSL within the generic example.
* Relocated enums and types to unique files.
* Added a header that includes types for the caliptra image.
  > Would prefer this to be generated via cbindgen from the image/types/
    rust library, but cbindgen wasn't having it.
* Fixed an issue in caliptra_fuse_array_write that indexed the array
  incorrectly, causing invalid data to be written to the right
  registers.
  > Fixed this in the c-binding's implementation as well, just for good
    measure.

* Relegate fuse setup to hwmodel

Per comments from reviewers, this commit both adds comments regarding
fuse setup and also relegates all digest work and fuse setup to the
hwmodel, as it requires it but actual hardware should not.

The swsim, verilator, and FPGA do not have true OTP fuses so they need
to have OTP fuse data programmed on startup by the host SOC. In an actual
device, the OTP fuse data would be programmed during manufacturing.

As such all fuse handling has been moved out of the generic/main.c and
out of the Caliptra C API proper.