Shrink ROM data section
Shrink the ROM data section to 4KiB. This prevents the compiler from
using large portions of DCCM to store ROM's .data and .bss regions.
This shouldn't impact ROM; as of this writing, those sections are all
empty in the caliptra-rom binary. This reorganization is just a
precaution to ensure changes aren't made to ROM which change this
situation.

This is to ensure FMC and RT have full control over a large portion of
DCCM (89KiB) which will be untouched by ROM. This is important because
it allows mutable firmware to store data in DCCM which will persist
across warm/update resets. Some of this region will be used to hold
runtime's DPE state (about 3.5 KiB).