emu-mbox: To match latest RTL, read 0 from fifo when empty.
When empty, reading from the fifo used to return the last item read from
the fifo, but that was changed in the RTL to return zeroes instead,
breaking the mbox driver tests when ran under verilator.

This commit changes the emulator and the driver tests to match the new
RTL behavior, and the tests now pass on both the emulated and verilated
models.