emu-mbox: Error when SoC clears exec bit unexpectedly.
The RTL used to ignore this, but now it transitions the FSM to the error
state. This PR tweaks the emulator to act the same as the RTL, and
modifies the test to expect this. Also, the test has been renamed and
expanded a bit to test what happens when the uC clears the error state.

The test now passes on both the emulated and verilated models.