Implement FIPS commands for ROM and invoke SELF-TEST from Runtime. (#693)
* Decouple dependency of Kat execution from Rom

* Decouple verification environment from ROM dependency

* Move fips-self-test test to test crate

* Use conditional compilation to build caliptra-runtime as ROM image

* Postpone self test execution

* Fix 1600 byte ROM size regression in #684

When callers from separate crates call a large generic function like
verify_lms_signature_cfi(), rustc 1.70 may build multiple versions
(depending on optimizer heuristics), even when all the generic
parameters are identical. This is bad, as it can bloat the binary and
the second copy violates the FIPS requirements that the same machine
code be used for the KAT as the actual implementation. To defend against
it, we provide a non-generic function that production firmware should
call instead.

* Rename goto_idle to enter_idle

* Fix typo

* Renamed SELF_TEST

* START/GET_PROGRESS commands

* kats executes before image verification

* Remove persistent_data from venv

* Implement FIPS commands in ROM

VERSION
SELF_TEST
SHUTDOWN

* Fix unassigned error codes

* Address feedback

- Remove mut from fips create_slice
- Use copy_bytes_to_mbox in write_response
- Restore idevid tests mistakenly removed
- Add bounds checking on fmc and rt sizes

* Rom itegrity test call from Runtime.

* Fixes for gate

---------

Co-authored-by: Anthony Rocha <anthony.rocha@amd.com>
Co-authored-by: Kor Nielsen <kor@google.com>