[fuzzing] Implementing initial support for fuzzing LMS (#618)
* treewide: Refactor SHA256 driver into an implemented trait

`Sha256HardwareDriver` is now the primary implementation of
`Sha256`. Where a concrete implementation is required,
`Sha256HardwareDriver` is almost always the correct instance.

This refactor makes it easier to test consumers of this driver, where
software implementations of SHA256 may be preferred. Follow-up commits
will provide a software implementation, which will then be used for
fuzz testing the LMS driver.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>

* [fuzzing] Implement initial support for LMS driver fuzzing

Instructions for running the supported fuzzers can be found at
https://gist.github.com/benjamindoron/59313b863b079f6f5af56af83d32648e.
This can be migrated to official documentation.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>

* Make Sha256 driver refactor as minimal as possible.

Signed-off-by: Benjamin Doron <benjamin.doron@9elements.com>
Co-authored-by: Kor Nielsen <kor@google.com>