UART driver: toggle a bit on every write.
This modification toggles generic-output-wire bit #8 every time a
character is written, so the outside world can tell when we've written a
new character without having to introspect internal signals.

As it is no longer necessary, remove the internal signal introspection
from the verilator hardware model and have it look at the new toggle bit
instead.