Move DPE Instance to persistent data in DCCM
This is necessary so that DPE measurements don't get cleared upon warm resets.
We will still need to add logic to RT to only initialize dpe upon
cold reset and validate the DpeInstance in DCCM, but these will
take longer and don't touch ROM so they can be done later.