Fix race when building same binary with different features.
To prevent a race condition with concurrent calls to caliptra-builder
from other threads or processes, we now hold a lock until we've read the
output binary from the filesystem (it's possible that another concurrent
process will build the same binary with different features before we get
a chance to read it).

I believe this race condition is responsible for intermittent test
failures of the FMC's test_fht_info() and test_pcr_log() tests in the
GHA environment.