[fix] TOCTOU Condition in SHA-512 Accelerator Lock Acquisition
This fix addreses https://github.com/chipsalliance/caliptra-sw/issues/841
On a Cold/Warm Reset, the SHA accelerator is guaranteed to be locked for Caliptra.
On an hitless update (aka Update Reset), it is expected that the SOC keeps the SHA Acc. unlocked.
Given these guarantees, this change addresses the issue by passing the lock state information to
the SHA Acc driver, which then acts on it by either attempting to acquire the SHA acc. lock or assuming it is already locked.