Validate DPE Instance in SRAM
In RT-fw, if we are not in a cold reset, we are expected to use
the DPE instance in SRAM. Before we use it, we need to validate
it to ensure that the DPE context tree is well-formed (the array
representing the contexts is actually a tree) and verify that
the current TCI measurement of the root TCI node matches the
latest runtime PCR value. Otherwise, we disable attestation.