rom-test: Fix race in test_fake_rom_update_reset()
The test FMC expects the command to be in the mailbox as soon as it
starts, which isn't possible due to the imprecise timing of the
fpga_realtime model. Switch the test to use the interactive test FMC
instead, and remove a potential race condition waiting for the
UpdateResetStarted boot status.