Consolidate runtime integration test binaries
This puts all the runtime integration tests into a single binary named
runtime_integration_tests-xxxx, which makes it much easier to run all the
tests on a remote system like a FPGA dev board.

This also has the side benefit of reducing compilation time and reducing
the total size of all test binaries, leading to faster transfers over
the network.

Note: Since all these tests are now in the same crate, and this crate
depends on serde_json, type interference doesn't know whether .into() is
referring to u32 or serde_json::Value. Thus, I've switched the code to
be explicit with u32::from.