Rearrange runtime test to fix timeout on FPGA.
caliptra-runtime::integration_tests::test_update failed because of a
timeout in the TRNG. The external TRNG is serviced by the AP cores in
the Zynq Ultrascale+. The test was previously booting Calitpra then
building the update image, but this would make the test fail because the
AP cores are too busy building to respond to the TRNG request. This
switches the order so it builds the update image and then starts booting
Caliptra. Which gives the AP cores the resources it needs to respond.