Add `fpga_realtime` feature to runtime. (#970)
This allows runtime to be built for the FPGA. There were several tests
that would fail on the FPGA because it was missing this feature. For
example

```bash
CPTRA_UIO_NUM=4 cargo nextest run --no-fail-fast --features=fpga_realtime --release -E 'package(caliptra-runtime)'
```

Would result in build errors because it wasn't able to specify the
`fpga_realtime` feature for `caliptra-runtime` in `caliptra-builder`.