tests: Produce FPGA-specific keyvault test binary
To save FPGA resources, the RTL used by the FPGA only has 16 key-vault
slots. The regular keyvault tests cannot pass on the FPGA because
they are checking all 32 slots. The test-case looks at the fpga_realtime
feature to decide how my slots to test.

Previously, caliptra-builder had its own fpga_realtime feature that was
being propagated to firmware targets, in particular this test. This
approach doesn't work when the firmware binaries are pre-built on
another machine (such as in CI), as the same instance of
caliptra-builder builds all the binaries, regardless of features.

This PR removes the fpga_realtime feature from caliptra_builder, and
instead registers different keyvault test binaries, one for the FPGA and
one for everything else, and decides from the host-side test logic which
one to run.