Update frozen images.
PR #1015 didn't update these. The policy check on that PR failed:

https://github.com/chipsalliance/caliptra-sw/actions/runs/6723263310/job/18272960952?pr=1015

Unfortunately, at the time that wasn't a required check, and didn't
block the auto-merge. That's rectified now in the repo settings.

Also, this commit runs the policy check on all pushes to main, to make
it clear in the commit history that it is passing at head.