Fix DPE HMAC bug where HMAC is always a zero array
The function hmac_sign_with_derived incorrectly computes an HMAC by
using an hmac key in the key vault and attempting to obtain an hmac
tag as an array. This is not allowed due to wanting to prevent
side-channel attacks.

With this change, we now derive an EC public key from the input CDI
and hash it to get ikm to the hmac kdf. This can then
be used to calculate an hmac value in array form.