Use sha2::Sha384 to hash in RT tests instead of mbox responder
Previously, we were using the sha384 driver via mbox responder
for hashing in RT tests since the sha2::Sha384 hash wasn't
matching what we were getting from the driver. I discovered
that this is due to the endianness being different.

Now, we can use sha2::Sha384 instead of mbox responder if we just
flip the endianness of the sha2::Sha384 hash.