Implement hash chain derivation in FMC.
This commit sets the max SVN value to 64. A hash chain of that length costs ~160k additional cycles to derive.

This commit also allocates reserved owner-signed bits for an epoch field, which is used when deriving the hash chain root. If the epoch is empty, no hash chain is derived; allows integrations that do not require this functionality to opt out.