ROM: Set ready_for_fw as early as possible.
Customers want to be able to stash measurements as early as possible.
Unfortunately, the existing ready_for_fw signal is sent much later than
necessary. Now that the mailbox SRAM is being initialized by the ROM,
customers MUST wait for ready_for_fw, so let's make it as painless as we
can.