runtime: Fixup NMI and race conditions in WFI logic
Previously, an unexpected NMI was firing with mcause=0xf0001002 (Machine Fast
Interrupt non-DCCM region), because the firmware wasn't setting up the
meivt CSR correctly.

Caliptra's VeeR instance has been configured with
RV_FAST_INTERRUPT_REDIRECT. This means that the regular risc-v external
interrupt dispatch is ignored, and instead the CPU is using a vector
table from the meivt register. The NMI fired because we weren't setting
this to a DCCM address.

This commit removes the hacks from the NMI handler, sets up a new
handler for external interrupts, and ensures that mie is only set when
the CPU is asleep, fixing the race condition.