sw-emulator mbox: Fix bug when datain is written before dlen.
The hardware is not fussy about the order in which dlen and datain
registers are used in the EXECUTE_UC state, but the sw-emulator was, and
this was causing a response in a ROM test to behave strangely (depending
on the exact contents of the mailbox request).

https://github.com/chipsalliance/caliptra-sw/blob/bcfef10cea57c586e27696288bc1899ccf065f3e/rom/dev/tools/test-fmc/src/main.rs#L318-L324

That test was writing to datain before setting dlen, which caused the
sw-emulator to return part of the request as the response instead of the
written response. The test was only passing by coincidence, as the
mailbox request happened to contain zeroes. And worse, the test wouldn't
catch real errors if they occurred.

This PR fixes the emulator to behave like the hardware, and adds a
hw-model test case to ensure the behavior is the same in the RTL and
sw-emulator.