Don't randomly init SRAM when testing lightweight "test ROMs"
To improve verilator times, caliptra-test-harness does not initialize
SRAM contents, so we must run our tests with the memory cleared to
zeroes, in case the compiler generates a speculative read to
uninitialized memory (this is rare, but tricky to debug when it
happens).

PR #1404 tweaks one of these test binaries enough that the compiler
generates a speculative read to uninitialized memory, which
intermittently triggers an ECC double-bit error.