cfi: Simpler launder implementation for common types
`cfi_launder` is meant to prevent the Rust compiler from optimizing a
value away.

Our current implementation uses `core::hint::black_box()`, which is the
recommended way in Rust. The problem is, this appears to often force the
argument to spill into memory and to be reloaded, which can be a lot of
extra instructions.

The original inspiration for this function is from, I believe, [OpenTitan's launder*
functions](https://github.com/lowRISC/opentitan/blob/master/sw/device/lib/base/hardened.h#L193).
There, they use an LLVM-specific trick of a blank inline assembly block
to force the compiler to keep the argument in a register.

After reviewing our code and speaking with @vsonims, it sounds like the
intention of the launder in our code is to prevent the compiler from
optimizing the value away (as the comments suggest), so the simpler
inline assembly trick may be sufficient (since we use the official
Rust compiler, which uses LLVM).

The biggest problem is that we launder many types of values in our code
and not all of them fit into a register.

So, this PR represents an incremental change: for `u32`s and similar
small types, we implement `cfi_launder` using the inline assembly trick
from OpenTitan. For any other types, we have a trait that can be derived
that will call `core::hint::black_box` in the same way as today.

We can do future follow-up PRs to try to try to clean up some of those
other uses of `cfi_launder` to hopefully shrink the code more.

This PR appears to shrink the ROM code size by 1152 bytes and the runtime
firmware by 616 bytes.