runtime docs: Remove storage from cryptographic mailbox
Mostly.

We need to keep a limited amount of storage per AES-256-GCM key to track
usage, since these keys can only be used a certain number of times due
to their IV problems.

For HMAC, we don't specify a key usage on the output MAC, since it has
to be imported back into Caliptra to be used as a key, where its usage
will have to be specified.

We also fix a few other minor issues with the mailbox documentation:

Fixes #1753
Fixes #1754
Fixes #1755
Fixes #1756