[ROM] Only inspect RT FW's SVN during image validation.
This is done as part of Stable Identity, which requires Caliptra
to have a single SVN. As part of this work, FMC's SVN is being
deprecated. Instead, the runtime SVN now describes the security
state of the entire firmware image. This is safe because the
FMC and runtime images are always signed, validated, and updated
in tandem.

Future changes will update build tooling to prevent callers from
setting FMC's SVN.