Update to rust 1.83
The goal of updating the rust library is to be able to use Mldsa from
rustCrypto repo: https://github.com/RustCrypto/signatures/pull/877

The goal of using this code is to replace openssl to generate TBS and
CSR tempaltes. Openssl does not yet support Mldsa

The minimum required rust version is 1.81.

It looks like this toolchain needs a lot more stack than 1.70.

Rust 1.83 warns about unused stuff a lot more often which is addressed
in this commit.

Signed-off-by: Arthur Heymans <arthur.heymans@9elements.com>