Implement key ladder in ROM and RT FW.
The key ladder is initialized in cold-boot from LDevID CDI. It is lengthened during
update-reset flows to match min-SVN-since-cold-boot.

This PR does not include a smoke test, because the production application firmware
doesn't yet include any APIs for exercising the key ladder.

The additional cycles from this feature depends on the firmware's SVN, and is
roughly 2.7k cycles * (128 - fw_svn). In the worst case, when fw_svn = 0, the
cycle count is ~350k.