emulator/imagegen: Update fips204 crate for deterministic signing (#1866)
Updates to the latest `fips204` crate, which added support for
deterministic signing (which the spec states we should do by passing
in a fixed "rnd" seed).

For key generation, the spec generally does not let you pass in a fixed
ξ seed (since it is supposed to be "fresh"), but an external ξ seed is
generally allowed. So, we add a simple playback RNG that will allow the
ξ seed to be replayed, which will match this mode.

The emulator should now reflect what the hardware will do, and hopefully
the test vectors will match what we will get when we run this on real
hardware.