Update image-builder to output a hashes file.
This can be useful for offline/external signing to extract the digests
for an image bundle to be passed to the signing interface.

The output is a JSON file with the vendor and owner hashes tagged with
their respective names.

This was tested by printing the hash from the ROM in the smoke tests to
ensure the calculated hash is the same for both the firmware and tool.

(cherry picked from commit 6f0a823e81401c1b7735cbb55252427d6d9ba0c1)