Use a chunk iterator instead of a while loop
Currently `rand_bytes` is only used on a buffer < than 48 bytes.

If used with a buffer larger than 48 bytes a panic is introduced to the
Caliptra runtime due to bounds checking.

This change uses a chunks iterator instead. Each chunk is guaranteed to
be 48 bytes except for the last chunk which contains the remainder.