Create an all zeros fw image
Creates an image bundle with FMC and Runtime populated with zeros. 0 is
an ilegal opcode in risc-v, so the image will immediately NMI.

This is useful for testing production signing infrastructure. A
production key can be used to sign this image before an official
release is ready. The prod-signed zeros binary is benign but can
be used to ensure that the prod key was loaded into fuses correctly.