Add fuse controller
This adds both the fuse controller asa the UDS fuses. To make things
easiers the UDS fuses are made part of the fuse controller with a
hardcoded offset. This is what the RDL suggest that this needs to be the
case as only the lowest 12 bits of the address gets used.

Signed-off-by: Arthur Heymans <arthur.heymans@9elements.com>