emulator: Add event bus between Caliptra, MCU, BMC, etc.
And hook it up to the emulated AXI bus for DMA transfers.

The events are used to simulate some of the ways that Caliptra
communicates with others in hardware, in particular, I3C, writing to
MCU's memory aperture, and the recovery communications.

We wire these up in the emulated AXI DMA controller and recovery
interface.

This PR will be coordinated with a PR on the MCU repository, which will
provide the emulated MCU and BMC cores, and will test the interactions.