image: Implement new methods for RustCrypto provider
Some new methods were added to the image crypto trait but the rustcrypto
provider was not updated, so the build was failing with rustcrypto.

Also, the rustcrypto provider was specifying compressed points for ECDSA
public keys, even though for OpenSSL we always use uncompressed, which
was causing keys not to be parsed correctly in RustCrypto.

To ensure that won't happen again, we also add a test that the workspace
members that use rustcrypto are now built with both sets of flags.

We also add an option to print the vendor and owner PK hashes when
building an firmware image bundle. This makes it easier to test the
recovery flow.