Update SoC Manifest and Mailbox Spec to support Hitless Updates
To support Hitless Updates, the following changes are needed

1. SoC Manifest
- Add staging address in the image metadata entry. This will be used by
MCU to temporarily store the image being downloaded through PLDM

2. Mailbox
- Update SET_AUTH_MANIFEST command to reflect all the fields in the SoC
Manifest IME
- Update AUTHORIZE_AND_STASH command to specify if the digest to be
computed will be coming from the load or staging area. This command will
need to DMA the contents of the load/staging area to the SHA accelerator
to compute the digest.
- Add GET_IME_COUNT - this command will inform the MCU how many firmware
images are in the SoC Manifest, so that it can query the metadata entry
one by one. This is needed to construct the FirmwareParameters PLDM
message to BMC.
- Add GET_IME - this command will retrieve a specific IME given an
index. This will be used by the firmware update and streaming boot
applications in the MCU
- Add ACTIVATE_FIRMWARE - this command activates the particular firmware
image. This is sent by MCU when it receives an 'ActivateFirmware'
message from PLDM.