Fix up image size like the emulator
The code used to be like this but the rework of #1904 changed that
again.

The RDL does not specify the semantics of the registers yet and the
weird multi dword semantics comes from an old version of OCP I3C
commands, which likely don't translate to how the hardware really works.

Signed-off-by: Arthur Heymans <arthur.heymans@9elements.com>