Update to zerocopy 0.8.17
Also updates DPE submodule to pull in the same zerocopy.

Zerocopy's derive macro for KnownLayout has changed. Compiling DPE in caliptra-sw fails because public interfaces are leaking these private types.

The solution is to mark them as a public field. There may be another workaround, but since these types are being serialized to and from raw bytes, it makes sense that these private fields should be visible to callers.