rom/runtime: Verify auth manifest in recovery flow
This uses the same code as the SET_AUTH_MANIFEST command, which has been
slightly refactored to accommodate reading directly from the entire
mailbox as an option.

Due to a timing issue when writing the FW index, the Caliptra FW (index
0) was getting sent instead of the SoC manifest (index 1). To account
for this, we made a small ROM change to pre-emptively set the
recovery FW index to 1 when it is finishing up.

I did some cleanup on the constant names in the recovery interface,
since it was not always clear which values were for the device status
and which were for the recovery status.

This was verified as working with the MCU emulator.

Next step is validating the MCU firmware, but this will be more complex
since we'll have to do some DMA gymnastics to ship the MCU SRAM
to the SHA accelerator, so we will do that in a separate PR.