docs: Clarify `ATMN` and `CMAN` magic markers
When debugging an issue, I noticed that the auth manifest magic marker
was appearing as `NMTA` on disk:

```shell
$ hexdump -C soc-manifest | head -n 1
00000000  4e 4d 54 41 00 1c 00 00  01 00 00 00 01 00 00 00  |NMTA............|
```

This appears consistent with the current documentation, which says that
the field should be the little-endian encoding of `0x41544D4E`.

However, this is inconsistent with how we interpret the FW magic marker,
which is specified as the little-endian encoding of `0x434D414E` (which
would be `NAMC`), but actually appears on disk as `CMAN`:

```shell
$ hexdump -C caliptra-fw-bundle.bin | head -n 1
00000000  43 4d 41 4e 3c 42 00 00  03 00 00 00 01 00 00 04  |CMAN<B..........|
```

The intent of the the auth manifest marker was probably to be `ATMN`,
but right now it is easier to change the spec than the code.

I updated the spec for both to be crystal clear about the order of the
bytes in both, and made them consistent with each other.